[REF] l10n_br_fiscal: extract l10n_br_fiscal_notification
