+++ l10n_br_fiscal/__manifest__.py
@@ -21,7 +21,6 @@
         "security/ir.model.access.csv",
         # Data
         # Some data is being loaded via post_init_hook in hook file
-        "data/l10n_br_fiscal_email_template.xml",
         "data/l10n_br_fiscal_data.xml",
         "data/uom_data.xml",
         "data/uom_alternative_data.xml",
@@ -62,7 +61,6 @@
         "views/product_genre_view.xml",
         "views/document_type_view.xml",
         "views/document_serie_view.xml",
-        "views/document_email_view.xml",
         "views/simplified_tax_view.xml",
         "views/simplified_tax_range_view.xml",
         "views/operation_view.xml",
