+++ l10n_br_fiscal_notification/data/l10n_br_fiscal_email_template.xml
