+++ l10n_br_fiscal_notification/models/__init__.py
@@ -1,0 +1,1 @@
+from . import document
